+++ .migration_backup/package.json
@@ -2,50 +2,54 @@
   "name": "tra-app",
   "version": "0.0.1",
   "description": "A native application built with NativeScript-Vue",
-  "author": "svend <sven.depickere@student.howest.be>",
+  "author": "wingcrowny@gmail.com",
   "license": "AGPL-3.0",
   "nativescript": {
     "id": "com.wingcrony.tra.app",
     "tns-ios": {
-      "version": "5.0.0"
+      "version": "6.4.1"
     },
     "tns-android": {
-      "version": "5.0.0"
+      "version": "6.4.1"
     }
   },
   "dependencies": {
-    "@vue/devtools": "^5.0.6",
-    "moment": "2.29.2",
-    "nativescript-feedback": "1.3.10",
-    "nativescript-socketio": "^3.2.1",
-    "nativescript-theme-core": "^1.0.4",
-    "nativescript-toasty": "^1.3.0",
-    "nativescript-vue": "^2.0.0",
+    "@nstudio/nativescript-loading-indicator": "^3.0.1",
+    "@vue/devtools": "^5.3.3",
+    "moment": "2.24.0",
+    "nativescript-feedback": "1.3.12",
+    "nativescript-gif": "4.0.2",
+    "nativescript-social-share": "1.5.2",
+    "nativescript-socketio": "^3.3.1",
+    "nativescript-theme-core": "^2.0.24",
+    "nativescript-toasty": "^2.0.1",
+    "nativescript-vue": "^2.5.0-alpha.3",
     "nativescript-vue-devtools": "^1.2.0",
-    "tns-core-modules": "^5.0.2",
-    "vuex": "^3.0.1"
+    "tns-core-modules": "~6.4.0",
+    "vuex": "^3.1.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.0.0",
-    "@babel/preset-env": "^7.0.0",
-    "babel-loader": "^8.0.2",
+    "@babel/core": "^7.7.7",
+    "@babel/preset-env": "^7.7.7",
+    "babel-loader": "^8.0.6",
     "babel-traverse": "6.26.0",
     "babel-types": "6.26.0",
     "babylon": "6.18.0",
-    "clean-webpack-plugin": "^0.1.19",
-    "copy-webpack-plugin": "^4.5.2",
-    "css-loader": "^1.0.0",
+    "clean-webpack-plugin": "^3.0.0",
+    "copy-webpack-plugin": "^5.1.1",
+    "css-loader": "^3.4.0",
     "lazy": "1.0.11",
-    "nativescript-dev-webpack": "0.24.1",
-    "nativescript-localstorage": "2.0.0",
-    "nativescript-vue-template-compiler": "^2.0.0",
-    "nativescript-worker-loader": "~0.9.0",
-    "node-sass": "^7.0.0",
-    "sass-loader": "^7.1.0",
-    "terser-webpack-plugin": "^1.1.0",
-    "vue-loader": "^15.2.6",
-    "webpack": "^4.16.4",
-    "webpack-bundle-analyzer": "~3.3.2",
-    "webpack-cli": "^3.1.0"
+    "nativescript-dev-webpack": "~1.4.0",
+    "nativescript-localstorage": "2.0.1",
+    "nativescript-vue-template-compiler": "^2.4.0",
+    "nativescript-worker-loader": "~0.10.0",
+    "node-sass": "^4.13.1",
+    "sass-loader": "^8.0.1",
+    "style-loader": "^1.1.3",
+    "terser-webpack-plugin": "^2.3.1",
+    "vue-loader": "^15.8.3",
+    "webpack": "^4.41.4",
+    "webpack-bundle-analyzer": "~3.6.0",
+    "webpack-cli": "^3.3.10"
   }
 }
