+++ nativescript.config.ts
@@ -1,0 +1,11 @@
+import { NativeScriptConfig } from '@nativescript/core'
+
+export default {
+  id: 'com.wingcrony.tra.app',
+  appPath: 'app',
+  appResourcesPath: 'app/App_Resources',
+  android: {
+    v8Flags: '--expose_gc',
+    markingMode: 'none',
+  },
+} as NativeScriptConfig
