+++ package.json
@@ -4,15 +4,6 @@
   "description": "A native application built with NativeScript-Vue",
   "author": "wingcrowny@gmail.com",
   "license": "AGPL-3.0",
-  "nativescript": {
-    "id": "com.wingcrony.tra.app",
-    "tns-ios": {
-      "version": "6.4.1"
-    },
-    "tns-android": {
-      "version": "6.4.1"
-    }
-  },
   "dependencies": {
     "@nstudio/nativescript-loading-indicator": "^3.0.1",
     "@vue/devtools": "^5.3.3",
@@ -25,8 +16,8 @@
     "nativescript-toasty": "^2.0.1",
     "nativescript-vue": "^2.5.0-alpha.3",
     "nativescript-vue-devtools": "^1.2.0",
-    "tns-core-modules": "~6.4.0",
-    "vuex": "^3.1.2"
+    "vuex": "^3.1.2",
+    "@nativescript/core": "~8.4.0"
   },
   "devDependencies": {
     "@babel/core": "^7.7.7",
@@ -35,21 +26,13 @@
     "babel-traverse": "6.26.0",
     "babel-types": "6.26.0",
     "babylon": "6.18.0",
-    "clean-webpack-plugin": "^3.0.0",
-    "copy-webpack-plugin": "^5.1.1",
-    "css-loader": "^3.4.0",
     "lazy": "1.0.11",
-    "nativescript-dev-webpack": "~1.4.0",
     "nativescript-localstorage": "2.0.1",
     "nativescript-vue-template-compiler": "^2.4.0",
-    "nativescript-worker-loader": "~0.10.0",
-    "node-sass": "^4.13.1",
-    "sass-loader": "^8.0.1",
     "style-loader": "^1.1.3",
-    "terser-webpack-plugin": "^2.3.1",
     "vue-loader": "^15.8.3",
-    "webpack": "^4.41.4",
-    "webpack-bundle-analyzer": "~3.6.0",
-    "webpack-cli": "^3.3.10"
-  }
+    "@nativescript/webpack": "~5.0.0",
+    "sass": "~1.49.9"
+  },
+  "main": "./app/main.js"
 }
